+++ programs/drift/src/controller/amm.rs
@@ -185,20 +185,33 @@ pub fn update_spreads(
     let max_ref_offset = market.amm.get_max_reference_price_offset()?;
 
     let reference_price_offset = if max_ref_offset > 0 {
-        let liquidity_ratio = amm_spread::calculate_inventory_liquidity_ratio(
-            market.amm.base_asset_amount_with_amm,
-            market.amm.base_asset_reserve,
-            market.amm.min_base_asset_reserve,
-            market.amm.max_base_asset_reserve,
-        )?;
+        let liquidity_ratio =
+            amm_spread::calculate_inventory_liquidity_ratio_for_reference_price_offset(
+                market.amm.base_asset_amount_with_amm,
+                market.amm.base_asset_reserve,
+                market.amm.min_base_asset_reserve,
+                market.amm.max_base_asset_reserve,
+            )?;
 
         let signed_liquidity_ratio =
             liquidity_ratio.safe_mul(market.amm.get_protocol_owned_position()?.signum().cast()?)?;
 
+        let deadband_pct = market.amm.get_reference_price_offset_deadband_pct()?;
+        let liquidity_fraction_after_deadband =
+            if signed_liquidity_ratio.unsigned_abs() <= deadband_pct {
+                0
+            } else {
+                signed_liquidity_ratio.safe_sub(
+                    deadband_pct
+                        .cast::<i128>()?
+                        .safe_mul(signed_liquidity_ratio.signum())?,
+                )?
+            };
+
         amm_spread::calculate_reference_price_offset(
             reserve_price,
             market.amm.last_24h_avg_funding_rate,
-            signed_liquidity_ratio,
+            liquidity_fraction_after_deadband,
             market.amm.min_order_size,
             market
                 .amm